refactor(cloud-functions): improve folder structure and logs
