+++ docs/logging_emojis.md
@@ -1,0 +1,52 @@
+# Unicode Emojis for Server Logs
+
+## Log Levels
+
+- ✅ **Success:** `✔️` (U+2714) / `✅` (U+2705)
+- ⚠️ **Warning:** `⚠️` (U+26A0)
+- ❌ **Error:** `❌` (U+274C) / `⛔` (U+26D4)
+- 🛑 **Critical:** `🛑` (U+1F6D1)
+- ℹ️ **Info:** `ℹ️` (U+2139)
+- 🔍 **Debug:** `🔍` (U+1F50D)
+- 📝 **Debug Information:** `📝` (U+1F4DD)
+
+## Request Handling
+
+- 📥 **Received Request:** `📥` (U+1F4E5)
+- 📤 **Sent Response:** `📤` (U+1F4E4)
+- ⏳ **Processing:** `⏳` (U+23F3)
+- 🕒 **Timeout:** `🕒` (U+1F552)
+- 🚫 **Blocked:** `🚫` (U+1F6AB)
+- 🔄 **Retrying:** `🔄` (U+1F504)
+
+## Database Events
+
+- 🗄️ **Database Query:** `🗄️` (U+1F5C4)
+- 🔍 **Cache Hit:** `🔍` (U+1F50D)
+- ❄️ **Cache Miss:** `❄️` (U+2744)
+- 💾 **Saved to DB:** `💾` (U+1F4BE)
+- 🚀 **Optimized Query:** `🚀` (U+1F680)
+
+## Security Events
+
+- 🔑 **Authentication:** `🔑` (U+1F511)
+- 🔓 **Access Granted:** `🔓` (U+1F513)
+- 🔒 **Access Denied:** `🔒` (U+1F512)
+- 🛡️ **Security Alert:** `🛡️` (U+1F6E1)
+- 👀 **Suspicious Activity:** `👀` (U+1F440)
+
+## System Events
+
+- 🖥️ **System Startup:** `🖥️` (U+1F5A5)
+- ⚡ **High CPU Usage:** `⚡` (U+26A1)
+- 🛠️ **Maintenance Mode:** `🛠️` (U+1F6E0)
+- 💣 **Crash/Error:** `💣` (U+1F4A3)
+- 🏗️ **Deployment:** `🏗️` (U+1F3D7)
+
+## Server Status
+
+- 🟢 **Online:** `🟢` (U+1F7E2)
+- 🔴 **Offline:** `🔴` (U+1F534)
+- 🟡 **Degraded:** `🟡` (U+1F7E1)
+- 🔄 **Restarting:** `🔄` (U+1F504)
+- 🚦 **Load Balancing:** `🚦` (U+1F6A6)
