+++ functions/src/core/extensions/math-extensions.ts
