+++ functions/src/core/extensions/number-extensions.ts
