+++ functions/src/core/extensions/string-extensions.ts
