+++ functions/src/shared/utils/ddos-ip-check.ts
@@ -4,9 +4,9 @@ import {
   Reference,
   ServerValue,
 } from 'firebase-admin/database';
-import { warn } from 'firebase-functions/logger';
+import { debug, warn } from 'firebase-functions/logger';
 import * as https from 'firebase-functions/v2/https';
-import { DATABASE_DEFAULT_URL } from './database_urls';
+import { DATABASE_DEFAULT_URL } from '../../core/constants/database-endpoint.constants';
 
 export const ddosCheck = async (
   req: https.Request,
@@ -14,7 +14,7 @@ export const ddosCheck = async (
   limit = 200,
 ): Promise<boolean> => {
   if (typeof limit !== 'number') {
-    warn('ddos value `limit` is not a number');
+    warn('⚠️ ddos value `limit` is not a number');
     limit = 200;
   }
 
@@ -30,22 +30,33 @@ export const ddosCheck = async (
     .child(`${dt.year}-${dt.month}-${dt.day}-${dt.hours}`);
 
   if (await isIpOnBlacklist(db, ipAddress)) {
+    warn(`🔒 Blocked IP ${ipAddress} tried to access again.`);
     return true;
   }
 
   const ipCount = await (await ddosRef.get()).val();
   if (ipCount > limit) {
     // DDOS Attack => block user
-    warn(`DDOS ATTACK BLOCKED | IP-Address: ${ipAddress}`);
-    await db.ref('blacklist').child(ipAddress).set(true).catch(warn);
+    warn(`🛡️ DDOS ATTACK BLOCKED | IP-Address: ${ipAddress}`);
+    await db
+      .ref('blacklist')
+      .child(ipAddress)
+      .set(true)
+      .catch((err) => {
+        warn(`❌ Failed to add ip ${ipAddress} to blacklist`, err);
+      });
     return true;
   } else if (!ipCount) {
     // Create new DDOS counter if there is no DDOS count
     await createNewDdosCounter({ db, counterName, ipAddress, ddosRef });
   } else {
     // Increment DDOS counter
-    await ddosRef.set(ServerValue.increment(1)).catch(warn);
+    await ddosRef.set(ServerValue.increment(1)).catch((err) => {
+      warn('❌ Failed to increment ddos counter', err);
+    });
   }
+
+  debug(`🔓 Passed DDOS check`);
 
   return false;
 };
@@ -62,9 +73,13 @@ const createNewDdosCounter = async (value: {
     .child(value.counterName)
     .child(value.ipAddress)
     .remove()
-    .catch(warn);
+    .catch((err) => {
+      warn('❌ Failed to remove old ddos counters', err);
+    });
   // Create new counter
-  await value.ddosRef.set(1).catch(warn);
+  await value.ddosRef.set(1).catch((err) => {
+    warn('❌ Failed to create ddos counter', err);
+  });
 };
 
 const extractIpAddress = (req: https.Request): string | null => {
